Rename 'allpatientsinfacility' package to 'allpatientsinfacility_old'
