+++ app/src/main/java/org/simple/clinic/activity/TheActivityComponent.kt
@@ -10,7 +10,7 @@ import dagger.Subcomponent
 import io.reactivex.Observable
 import org.simple.clinic.addidtopatient.searchforpatient.AddIdToPatientSearchScreen
 import org.simple.clinic.addidtopatient.searchresults.AddIdToPatientSearchResultsScreen
-import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityView
+import org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilityView
 import org.simple.clinic.bp.entry.BloodPressureEntrySheet
 import org.simple.clinic.bp.entry.ConfirmRemoveBloodPressureDialog
 import org.simple.clinic.bp.entry.di.BloodPressureEntryModule
