--- app/src/main/java/org/simple/clinic/allpatientsinfacility/AllPatientsInFacilityResultsAdapter.kt
@@ -1,3 +1,0 @@
-package org.simple.clinic.allpatientsinfacility
-
-import androidx.recyclerview.widget.ListAdapter
