+++ app/src/main/java/org/simple/clinic/allpatientsinfacility_old/AllPatientsInFacilityUi.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.allpatientsinfacility
+package org.simple.clinic.allpatientsinfacility_old
 
 interface AllPatientsInFacilityUi {
