+++ app/src/main/java/org/simple/clinic/allpatientsinfacility_old/AllPatientsInFacilityUiEvents.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.allpatientsinfacility
+package org.simple.clinic.allpatientsinfacility_old
 
 import org.simple.clinic.widgets.UiEvent
 import java.util.UUID
