+++ app/src/main/java/org/simple/clinic/search/PatientSearchScreen.kt
@@ -12,9 +12,9 @@ import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_patient_search.view.*
 import org.simple.clinic.R
 import org.simple.clinic.activity.TheActivity
-import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityListScrolled
-import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilitySearchResultClicked
-import org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityView
+import org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilityListScrolled
+import org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilitySearchResultClicked
+import org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilityView
 import org.simple.clinic.bindUiToController
 import org.simple.clinic.patient.PatientSearchCriteria
 import org.simple.clinic.router.screen.ScreenRouter
