+++ app/src/main/res/layout/view_allpatientsinfacility.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityView xmlns:android="http://schemas.android.com/apk/res/android"
+<org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilityView xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:app="http://schemas.android.com/apk/res-auto"
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
@@ -48,4 +48,4 @@
     app:fastScrollTrackColor="@color/transparent"
     tools:listitem="@layout/list_patient_search" />
 
-</org.simple.clinic.allpatientsinfacility.AllPatientsInFacilityView>
+</org.simple.clinic.allpatientsinfacility_old.AllPatientsInFacilityView>
