+++ app/src/test/java/org/simple/clinic/allpatientsinfacility_old/AllPatientsInFacilityUiStateProducerTest.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.allpatientsinfacility
+package org.simple.clinic.allpatientsinfacility_old
 
 import com.nhaarman.mockito_kotlin.mock
 import com.nhaarman.mockito_kotlin.whenever
